rename: modalSlice로 네이밍 변경
