+++ src/redux/rootReducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from "redux";
 import categorySlice from "./categorySlice.js";
-import modalSlice from "./modalState.js";
+import modalSlice from "./modalSlice.js";
 import selectedRestaurantItemSlice from "./selectedRestaurantItemSlice.js";
 import restaurantsSlice from "./restaurantsSlice.js";
